requirements.txt
flask
pillow
pycups